# docs/swagger.yaml
basePath: /
definitions:
  core.ApiOKResponse:
    properties:
      code:
        description: "200"
        type: integer
      data:
        type: string
      message:
        description: '"success"'
        type: string
    type: object
...
host: localhost:8060
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: go subpub service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Swagger Example API
  version: "2.0"
paths:
...
  /api/v1/publish/{topic}:
    post:
      description: Publish a message to a topic
      parameters:
      - description: Topic name
        in: path
        name: topic
        required: true
        type: string
      - description: Message to publish
        in: query
        name: message
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ApiOKResponse'
      summary: Publish a message to a topic
      tags:
      - Publish
...
schemes:
- http
swagger: "2.0"